Restructure command packages and add NPC functionality to renamepet
