+++ core/src/main/java/be/isach/ultracosmetics/UltraCosmetics.java
@@ -1,7 +1,7 @@
 package be.isach.ultracosmetics;
 
-import be.isach.ultracosmetics.command.CommandManager;
-import be.isach.ultracosmetics.command.CommandShowcaseManager;
+import be.isach.ultracosmetics.command.ultracosmetics.CommandManager;
+import be.isach.ultracosmetics.command.showcase.CommandShowcaseManager;
 import be.isach.ultracosmetics.config.MessageManager;
 import be.isach.ultracosmetics.config.NPCManager;
 import be.isach.ultracosmetics.config.SettingsManager;
@@ -494,7 +494,7 @@ public class UltraCosmetics extends JavaPlugin {
     /**
      * @return Showcase Command Manager.
      */
-    public CommandShowcaseManager getCommandShowcaseManager() {
+    public CommandShowcaseManager getShowcaseCommandManager() {
         return commandShowcaseManager;
     }
 }
