+++ core/src/main/java/be/isach/ultracosmetics/command/SubCommand.java
@@ -80,7 +80,7 @@ public abstract class SubCommand implements ITabCompletable {
      * @param sender The player who executed the command.
      * @param args   The args of the command. (Includes the subcommand alias).
      */
-    protected abstract void onExePlayer(Player sender, String... args);
+    public abstract void onExePlayer(Player sender, String... args);
 
     /**
      * Called when the sub command is executed by console.
@@ -88,7 +88,7 @@ public abstract class SubCommand implements ITabCompletable {
      * @param sender The console sender who executed the command.
      * @param args   The args of the command. (Includes the subcommand alias).
      */
-    protected abstract void onExeConsole(ConsoleCommandSender sender, String... args);
+    public abstract void onExeConsole(ConsoleCommandSender sender, String... args);
 
     /**
      * Sent when player doesn't have permission to the command.
